# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "File Management System API",
    "description": "Backend for a complete file management system: JWT auth with email-verification OTPs, file uploads with content extraction (PDF/DOCX/XLSX/CSV/text/images) and Cloudinary storage, per-user dashboards, admin statistics, and admin user management.\n\nAll authenticated endpoints expect `Authorization: Bearer <accessToken>`. The token payload carries `{ userId, role }` only.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" },
    { "url": "https://your-project.vercel.app", "description": "Vercel deployment (replace with your domain)" }
  ],
...
  "paths": {
...
    "/files/upload": {
      "post": {
        "tags": ["Files"],
        "summary": "Upload files (batch)",
        "description": "Authenticated multipart upload, field name `files` (up to 10 files; the WHOLE batch must total ≤4MB). Pipeline per file: memoryStorage (never disk) → backend derives the real MIME type from magic bytes (client-declared type is only a hint for undetectable text formats) → allow-list check → content extraction ∥ Cloudinary upload in parallel → File row persisted.\n\nAllowed types: PDF, DOCX, XLS/XLSX, CSV, TXT, JPEG/PNG/WEBP/GIF. Every file is processed independently — one failure never blocks the rest. Extraction failures still persist that row as `extractionStatus: FAILED` while the upload itself succeeds.\n\nStatus codes: `201` all uploaded · `207` partial success (mixed results) · `400` nothing persisted due to client errors · `502` nothing persisted due to storage failure.",
        "operationId": "uploadFiles",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["files"],
                "properties": {
                  "files": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 10,
                    "items": { "type": "string", "format": "binary" },
                    "description": "Files to upload (batch total ≤4MB, allowed types listed above)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "All files uploaded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadBatchResponse" },
                "example": {
                  "message": "All 2 file(s) uploaded successfully",
                  "summary": { "total": 2, "uploaded": 2, "failed": 0 },
                  "results": [
                    {
                      "filename": "quarterly-report.pdf",
                      "success": true,
                      "file": {
                        "id": "3f1d2a94-6b7e-4f0a-9d21-8c4b91aa77e1",
                        "filename": "quarterly-report.pdf",
                        "mimeType": "application/pdf",
                        "size": 48213,
                        "url": "https://res.cloudinary.com/demo/raw/upload/v1787612345/file-management/9b8c1be2-07ba-4075-8c07-82e49ac505ea/kqzvbmwpzr5fsk1hjdxo.pdf",
                        "storageKey": "file-management/9b8c1be2-07ba-4075-8c07-82e49ac505ea/kqzvbmwpzr5fsk1hjdxo",
                        "extractedContent": "Quarterly revenue grew 12% across all regions...",
                        "extractedMetadata": { "pageCount": 4, "mimeDetectionMethod": "magic-bytes" },
                        "extractionStatus": "COMPLETED",
                        "uploadedById": "9b8c1be2-07ba-4075-8c07-82e49ac505ea",
                        "createdAt": "2026-08-25T09:12:45.120Z",
                        "updatedAt": "2026-08-25T09:12:47.881Z"
                      }
                    },
                    {
                      "filename": "notes.txt",
                      "success": true,
                      "file": {
                        "id": "a41d9f60-2c85-47ab-b3f2-7d90cc31ee04",
                        "filename": "notes.txt",
                        "mimeType": "text/plain",
                        "size": 29,
                        "url": "https://res.cloudinary.com/demo/raw/upload/v1787612346/file-management/9b8c1be2-07ba-4075-8c07-82e49ac505ea/mxktrwqylp8vbn2ecjha.txt",
                        "storageKey": "file-management/9b8c1be2-07ba-4075-8c07-82e49ac505ea/mxktrwqylp8vbn2ecjha",
                        "extractedContent": "Meeting notes from Tuesday...",
                        "extractedMetadata": { "charCount": 29, "lineCount": 1, "mimeDetectionMethod": "text-sniffing" },
                        "extractionStatus": "COMPLETED",
                        "uploadedById": "9b8c1be2-07ba-4075-8c07-82e49ac505ea",
                        "createdAt": "2026-08-25T09:12:46.010Z",
                        "updatedAt": "2026-08-25T09:12:47.120Z"
                      }
                    }
                  ]
                }
              }
            }
          },
          "207": {
            "description": "Partial success — some files uploaded, some failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadBatchResponse" },
                "example": {
                  "message": "Uploaded 2 of 3 files",
                  "summary": { "total": 3, "uploaded": 2, "failed": 1 },
                  "results": [
                    {
                      "filename": "quarterly-report.pdf",
                      "success": true,
                      "file": { "id": "3f1d2a94-6b7e-4f0a-9d21-8c4b91aa77e1", "filename": "quarterly-report.pdf", "mimeType": "application/pdf", "size": 48213, "extractionStatus": "COMPLETED" }
                    },
                    {
                      "filename": "archive.zip",
                      "success": false,
                      "error": { "code": "VALIDATION_ERROR", "message": "File type not allowed" }
                    },
                    {
                      "filename": "broken.docx",
                      "success": false,
                      "error": { "code": "INTERNAL_SERVER_ERROR", "message": "Something went wrong" }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Nothing persisted — empty request or batch total exceeds 4MB",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "examples": {
                  "emptyRequest": { "value": { "error": { "code": "VALIDATION_ERROR", "message": "At least one file is required (multipart field \"files\")" } } },
                  "totalTooLarge": { "value": { "error": { "code": "VALIDATION_ERROR", "message": "Total size of all uploaded files exceeds the maximum allowed request size" } } },
                  "allDisallowed": { "value": { "error": { "code": "VALIDATION_ERROR", "message": "File type not allowed" } } }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": {
            "description": "Cloudinary upload failed for every file",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "UPLOAD_FAILED", "message": "Cloud storage upload failed" } }
              }
            }
          }
        }
      }
    },
    "/files": {
      "get": {
        "tags": ["Files"],
        "summary": "List own files (search / filter / sort / paginate)",
        "description": "Authenticated. Returns only the caller's own uploads.\n\n`type` accepts an exact MIME (`application/pdf`) or a family (`image`, `image/*`). `sortBy` whitelist: createdAt, updatedAt, size, filename, mimeType.",
        "operationId": "listFiles",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 }, "example": 1 },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 }, "example": 10 },
          { "name": "search", "in": "query", "schema": { "type": "string", "minLength": 1, "maxLength": 255 }, "description": "Case-insensitive substring match on filename", "example": "report" },
          { "name": "type", "in": "query", "schema": { "type": "string", "minLength": 1, "maxLength": 100 }, "description": "Exact mime or family prefix", "example": "image/*" },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" }, "description": "createdAt >= from", "example": "2026-08-01T00:00:00.000Z" },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" }, "description": "createdAt <= to", "example": "2026-08-31T23:59:59.000Z" },
          { "name": "minSize", "in": "query", "schema": { "type": "integer", "minimum": 0 }, "description": "size >= minSize (bytes)", "example": 1024 },
          { "name": "maxSize", "in": "query", "schema": { "type": "integer", "minimum": 0 }, "description": "size <= maxSize (bytes)", "example": 10485760 },
          {
            "name": "sortBy",
            "in": "query",
            "schema": { "type": "string", "enum": ["createdAt", "updatedAt", "size", "filename", "mimeType"], "default": "createdAt" },
            "example": "createdAt"
          },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }, "example": "desc" }
        ],
        "responses": {
          "200": {
            "description": "Paginated file list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FileListResult" },
                "example": {
                  "data": [
                    {
                      "id": "3f1d2a94-6b7e-4f0a-9d21-8c4b91aa77e1",
                      "filename": "quarterly-report.pdf",
                      "mimeType": "application/pdf",
                      "size": 48213,
                      "url": "https://res.cloudinary.com/demo/raw/upload/v1787612345/file-management/9b8c1be2-07ba-4075-8c07-82e49ac505ea/kqzvbmwpzr5fsk1hjdxo.pdf",
                      "storageKey": "file-management/9b8c1be2-07ba-4075-8c07-82e49ac505ea/kqzvbmwpzr5fsk1hjdxo",
                      "extractedContent": "Quarterly revenue grew 12% across all regions...",
                      "extractedMetadata": { "pageCount": 4, "mimeDetectionMethod": "magic-bytes" },
                      "extractionStatus": "COMPLETED",
                      "uploadedById": "9b8c1be2-07ba-4075-8c07-82e49ac505ea",
                      "createdAt": "2026-08-25T09:12:45.120Z",
                      "updatedAt": "2026-08-25T09:12:47.881Z"
                    }
                  ],
                  "pagination": { "page": 1, "limit": 10, "total": 1, "totalPages": 1 }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter (unknown sortBy, limit > 100, inverted ranges, …)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Invalid request payload",
                    "details": [{ "path": "sortBy", "message": "Invalid enum value. Expected 'createdAt' | 'updatedAt' | 'size' | 'filename' | 'mimeType'" }]
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/files/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "example": "3f1d2a94-6b7e-4f0a-9d21-8c4b91aa77e1" }
      ],
      "get": {
        "tags": ["Files"],
        "summary": "Get file detail",
        "description": "Owner or admin. Non-owners receive an indistinguishable 404.",
        "operationId": "getFileById",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "File detail incl. extracted content",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FileResponse" },
                "example": {
                  "file": {
                    "id": "3f1d2a94-6b7e-4f0a-9d21-8c4b91aa77e1",
                    "filename": "quarterly-report.pdf",
                    "mimeType": "application/pdf",
                    "size": 48213,
                    "url": "https://res.cloudinary.com/demo/raw/upload/v1787612345/file-management/9b8c1be2-07ba-4075-8c07-82e49ac505ea/kqzvbmwpzr5fsk1hjdxo.pdf",
                    "storageKey": "file-management/9b8c1be2-07ba-4075-8c07-82e49ac505ea/kqzvbmwpzr5fsk1hjdxo",
                    "extractedContent": "Quarterly revenue grew 12% across all regions...",
                    "extractedMetadata": { "pageCount": 4, "mimeDetectionMethod": "magic-bytes" },
                    "extractionStatus": "COMPLETED",
                    "uploadedById": "9b8c1be2-07ba-4075-8c07-82e49ac505ea",
                    "createdAt": "2026-08-25T09:12:45.120Z",
                    "updatedAt": "2026-08-25T09:12:47.881Z"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/FileNotFound" }
        }
      },
      "patch": {
        "tags": ["Files"],
        "summary": "Rename a file",
        "description": "Owner or admin. Filename is trimmed/sanitized like at upload time.",
        "operationId": "renameFile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RenameInput" },
              "example": { "filename": "quarterly-report-final.pdf" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Renamed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FileMutationResponse" },
                "example": {
                  "message": "File renamed successfully",
                  "file": {
                    "id": "3f1d2a94-6b7e-4f0a-9d21-8c4b91aa77e1",
                    "filename": "quarterly-report-final.pdf",
                    "mimeType": "application/pdf",
                    "size": 48213,
                    "url": "https://res.cloudinary.com/demo/raw/upload/v1787612345/file-management/9b8c1be2-07ba-4075-8c07-82e49ac505ea/kqzvbmwpzr5fsk1hjdxo.pdf",
                    "storageKey": "file-management/9b8c1be2-07ba-4075-8c07-82e49ac505ea/kqzvbmwpzr5fsk1hjdxo",
                    "extractedContent": "Quarterly revenue grew 12% across all regions...",
                    "extractedMetadata": { "pageCount": 4, "mimeDetectionMethod": "magic-bytes" },
                    "extractionStatus": "COMPLETED",
                    "uploadedById": "9b8c1be2-07ba-4075-8c07-82e49ac505ea",
                    "createdAt": "2026-08-25T09:12:45.120Z",
                    "updatedAt": "2026-08-25T09:14:02.553Z"
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/FileNotFound" }
        }
      },
      "delete": {
        "tags": ["Files"],
        "summary": "Delete a file",
        "description": "Owner or admin. Removes the DB row and best-effort destroys the Cloudinary asset (image vs raw type-aware). Storage hiccups never fail an already-committed deletion.",
        "operationId": "deleteFile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageOnly" },
                "example": { "message": "File deleted successfully" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/FileNotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token returned by POST /auth/login. Payload: { userId, role }."
      }
    },
    "schemas": {
...
      "Role": { "type": "string", "enum": ["USER", "ADMIN"] },
      "ExtractionStatus": { "type": "string", "enum": ["PENDING", "PROCESSING", "COMPLETED", "FAILED"] },
      "MimeDetectionMethod": { "type": "string", "enum": ["magic-bytes", "ole2-signature", "text-sniffing"], "description": "How the stored mimeType was derived (audit trail)." },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "email": { "type": "string", "format": "email" },
          "role": { "$ref": "#/components/schemas/Role" },
          "isVerified": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "email", "role", "isVerified", "createdAt", "updatedAt"]
      },
...
      "RenameInput": {
        "type": "object",
        "required": ["filename"],
        "properties": { "filename": { "type": "string", "minLength": 1, "maxLength": 255 } }
      },
...
      "MessageOnly": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"]
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "minimum": 1 },
          "limit": { "type": "integer", "minimum": 1, "maximum": 100 },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer", "minimum": 1 }
        },
        "required": ["page", "limit", "total", "totalPages"]
      },
...
      "File": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "filename": { "type": "string", "maxLength": 255 },
          "mimeType": { "type": "string", "description": "Derived from content (magic bytes / OLE2 signature / text sniffing) — never trusted from the client." },
          "size": { "type": "integer", "description": "Bytes, derived server-side from the received buffer." },
          "url": { "type": "string", "format": "uri", "description": "Cloudinary secure URL." },
          "storageKey": { "type": "string", "description": "Cloudinary public_id." },
          "extractedContent": { "type": "string", "nullable": true, "description": "Extracted text (≤100k chars); null when nothing extractable or extraction failed." },
          "extractedMetadata": {
            "type": "object",
            "nullable": true,
            "description": "Extractor-specific metadata plus mimeDetectionMethod audit field.",
            "properties": {
              "mimeDetectionMethod": { "$ref": "#/components/schemas/MimeDetectionMethod" }
            },
            "additionalProperties": true
          },
          "extractionStatus": { "$ref": "#/components/schemas/ExtractionStatus" },
          "uploadedById": { "type": "string", "format": "uuid" },
          "uploadedBy": { "$ref": "#/components/schemas/User", "description": "Owner of the file (populated via relation)." },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "filename", "mimeType", "size", "url", "storageKey", "extractedContent", "extractedMetadata", "extractionStatus", "uploadedById", "createdAt", "updatedAt"]
      },
      "FileResponse": {
        "type": "object",
        "properties": { "file": { "$ref": "#/components/schemas/File" } },
        "required": ["file"]
      },
      "FileMutationResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "file": { "$ref": "#/components/schemas/File" }
        },
        "required": ["message", "file"]
      },
      "UploadSummary": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "minimum": 1, "maximum": 10 },
          "uploaded": { "type": "integer", "minimum": 0 },
          "failed": { "type": "integer", "minimum": 0 }
        },
        "required": ["total", "uploaded", "failed"]
      },
      "FileUploadSuccessEntry": {
        "type": "object",
        "properties": {
          "filename": { "type": "string" },
          "success": { "type": "boolean", "enum": [true] },
          "file": { "$ref": "#/components/schemas/File" }
        },
        "required": ["filename", "success", "file"]
      },
      "FileUploadFailedEntry": {
        "type": "object",
        "properties": {
          "filename": { "type": "string" },
          "success": { "type": "boolean", "enum": [false] },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["filename", "success", "error"]
      },
      "FileUploadResultEntry": {
        "oneOf": [
          { "$ref": "#/components/schemas/FileUploadSuccessEntry" },
          { "$ref": "#/components/schemas/FileUploadFailedEntry" }
        ]
      },
      "UploadBatchResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "summary": { "$ref": "#/components/schemas/UploadSummary" },
          "results": { "type": "array", "items": { "$ref": "#/components/schemas/FileUploadResultEntry" } }
        },
        "required": ["message", "summary", "results"]
      },
      "FileListResult": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/File" } },
          "pagination": { "$ref": "#/components/schemas/PaginationMeta" }
        },
        "required": ["data", "pagination"]
      },
...
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "path": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["path", "message"]
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": ["VALIDATION_ERROR", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "CONFLICT", "FILE_TOO_LARGE", "MULTIPART_ERROR", "UPLOAD_FAILED", "EMAIL_SEND_FAILED", "INTERNAL_SERVER_ERROR"]
              },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "nullable": true,
                "items": { "$ref": "#/components/schemas/ErrorDetail" },
                "description": "Present only for zod validation failures."
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request body failed validation",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Invalid request payload",
                "details": [
                  { "path": "name", "message": "Name is required" },
                  { "path": "email", "message": "Must be a valid email address" },
                  { "path": "password", "message": "Password must be at least 8 characters" }
                ]
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, invalid, or expired bearer token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": { "error": { "code": "UNAUTHORIZED", "message": "Authentication required" } }
          }
        }
      },
...
      "FileNotFound": {
        "description": "File missing or owned by someone else (indistinguishable 404)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": { "error": { "code": "NOT_FOUND", "message": "File not found" } }
          }
        }
      }
    }
  }
}